Restructure /src
