+++ src/components/layout-toggle/layout-toggle.scss
@@ -1,4 +1,4 @@
-@import "./variables";
+@import '../../styles/variables';
 
 .layout-toggle {
   display: flex;
