+++ src/components/note/NotePreview.jsx
