+++ src/components/note/note-header.scss
@@ -1,4 +1,4 @@
-@import 'variables.scss';
+@import '../../styles/variables';
 
 .note-header {
   background: $background-color-dark;
