+++ src/components/note/note.scss
@@ -1,5 +1,5 @@
-@import 'variables.scss';
-@import 'webkit-scrollbar-styles.scss';
+@import '../../styles/variables';
+@import '../../styles/webkit-scrollbar-styles';
 
 .selected-note {
   bottom: 0;
