+++ src/components/notes-list/CategorySelector.jsx
