+++ src/components/notes-list/notes-list.scss
@@ -1,5 +1,5 @@
-@import 'variables.scss';
-@import 'webkit-scrollbar-styles.scss';
+@import '../../styles/variables';
+@import '../../styles/webkit-scrollbar-styles';
 
 $header-height: 130px;
 $footer-height: 45px;
